llava-requirements.txt
langchain==0.0.351
hydra-core==1.3.2
open-clip-torch==2.24.0
datasets==2.16.1
netcal==1.3.5
uncertainty-calibration==0.1.4
seaborn==0.13.2